# public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"Field": {
				"properties": {
					"label": {
						"type": "string"
					},
					"value": {
						"type": "string"
					},
					"key": {
						"type": "string"
					}
				},
				"required": [
					"label",
					"value",
					"key"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Error": {
				"properties": {
					"message": {
						"type": "string"
					},
					"field": {
						"$ref": "#/components/schemas/Field"
					}
				},
				"required": [
					"message",
					"field"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ValidationErrorResponse": {
				"properties": {
					"name": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"status": {
						"type": "boolean"
					},
					"errors": {
						"items": {
							"$ref": "#/components/schemas/Error"
						},
						"type": "array"
					}
				},
				"required": [
					"name",
					"message",
					"status",
					"errors"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IRoomInputDTO": {
				"properties": {
					"name": {
						"type": "string"
					},
					"floor": {
						"type": "number",
						"format": "double"
					},
					"capacity": {
						"type": "number",
						"format": "double"
					},
					"assets": {
						"properties": {
							"parkingAndFacilities": {
								"properties": {
									"paidParkingGarageOffPremises": {
										"type": "boolean"
									},
									"paidParkingGarageOnPremises": {
										"type": "boolean"
									},
									"evCharger": {
										"type": "boolean"
									},
									"privateGym": {
										"type": "boolean"
									},
									"gym": {
										"type": "boolean"
									},
									"elevator": {
										"type": "boolean"
									}
								},
								"type": "object"
							},
							"entertainment": {
								"properties": {
									"exerciseEquipment": {
										"type": "boolean"
									},
									"bluetoothSoundSystem": {
										"type": "boolean"
									},
									"booksAndReadingMaterial": {
										"type": "boolean"
									},
									"gameConsole": {
										"type": "boolean"
									},
									"tv": {
										"type": "boolean"
									}
								},
								"type": "object"
							},
							"internetAndOffice": {
								"properties": {
									"dedicatedWorkspace": {
										"type": "boolean"
									},
									"wiFi": {
										"type": "boolean"
									}
								},
								"type": "object"
							},
							"bedroomAndLaundry": {
								"properties": {
									"clothingStorage": {
										"type": "boolean"
									},
									"iron": {
										"type": "boolean"
									},
									"roomDarkeningShades": {
										"type": "boolean"
									},
									"extraPillowsAndBlankets": {
										"type": "boolean"
									},
									"bedLinens": {
										"type": "boolean"
									},
									"hangers": {
										"type": "boolean"
									},
									"essentials": {
										"type": "boolean"
									},
									"dryer": {
										"type": "boolean"
									},
									"washer": {
										"type": "boolean"
									}
								},
								"type": "object"
							},
							"heatingAndCooling": {
								"properties": {
									"radiantHeating": {
										"type": "boolean"
									},
									"ceilingFan": {
										"type": "boolean"
									},
									"indoorFireplace": {
										"type": "boolean"
									},
									"heating": {
										"type": "boolean"
									},
									"portableFans": {
										"type": "boolean"
									},
									"airConditioning": {
										"type": "boolean"
									}
								},
								"type": "object"
							},
							"bathroom": {
								"properties": {
									"cleaningProducts": {
										"type": "boolean"
									},
									"showerGel": {
										"type": "boolean"
									},
									"hotWater": {
										"type": "boolean"
									},
									"bodySoap": {
										"type": "boolean"
									},
									"conditioner": {
										"type": "boolean"
									},
									"shampoo": {
										"type": "boolean"
									},
									"hairDryer": {
										"type": "boolean"
									},
									"bathtub": {
										"type": "boolean"
									}
								},
								"type": "object"
							},
							"kitchenAndDining": {
								"properties": {
									"barbecueUtensils": {
										"type": "boolean"
									},
									"oven": {
										"type": "boolean"
									},
									"cookingBasics": {
										"type": "boolean"
									},
									"diningTable": {
										"type": "boolean"
									},
									"trashCompactor": {
										"type": "boolean"
									},
									"toaster": {
										"type": "boolean"
									},
									"wineGlasses": {
										"type": "boolean"
									},
									"coffeeMaker": {
										"type": "boolean"
									},
									"hotWaterKettle": {
										"type": "boolean"
									},
									"stove": {
										"type": "boolean"
									},
									"dishwasher": {
										"type": "boolean"
									},
									"freezer": {
										"type": "boolean"
									},
									"dishesAndSilverware": {
										"type": "boolean"
									},
									"microwave": {
										"type": "boolean"
									},
									"refrigerator": {
										"type": "boolean"
									},
									"kitchen": {
										"type": "boolean"
									}
								},
								"type": "object"
							}
						},
						"type": "object"
					}
				},
				"required": [
					"name",
					"floor"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IUserInputDTO": {
				"properties": {
					"name": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"repeatPassword": {
						"type": "string"
					},
					"acceptTerms": {
						"type": "boolean"
					}
				},
				"required": [
					"name",
					"username",
					"email",
					"password",
					"repeatPassword",
					"acceptTerms"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ITokenInputDTO": {
				"properties": {
					"token": {
						"type": "string"
					}
				},
				"required": [
					"token"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LoggedUserResponse": {
				"properties": {
					"auth": {
						"type": "boolean"
					},
					"jwtToken": {
						"type": "string"
					},
					"refreshToken": {
						"type": "string"
					}
				},
				"required": [
					"auth",
					"jwtToken",
					"refreshToken"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IUserLoginDTO": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"description": "JWT Authorization header using the Bearer scheme.\n\nEnter your token in the text input below.\n\n Example: \"12345abcdef\"\n\n Name: Authorization\n\nIn: header"
			}
		}
	},
	"info": {
		"title": "booking-room-app",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/room/add": {
			"post": {
				"operationId": "AddNewRoom",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"422": {
						"description": "Validation Failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidationErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"name": "Validation Error.",
											"message": "Some fields are not valid.",
											"status": false,
											"errors": [
												{
													"message": "name is not allowed to be empty",
													"field": {
														"label": "name",
														"value": "",
														"key": "name"
													}
												}
											]
										}
									}
								}
							}
						}
					}
				},
				"tags": [
					"Rooms"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IRoomInputDTO"
							}
						}
					}
				}
			}
		},
		"/auth/register": {
			"post": {
				"operationId": "Register",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidationErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"name": "Validation Error.",
											"message": "Some fields are not valid.",
											"status": false,
											"errors": [
												{
													"message": "name is not allowed to be empty",
													"field": {
														"label": "name",
														"value": "",
														"key": "name"
													}
												}
											]
										}
									}
								}
							}
						}
					}
				},
				"tags": [
					"Authorization"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUserInputDTO"
							}
						}
					}
				}
			}
		},
		"/auth/verify-email": {
			"post": {
				"operationId": "VerifyAccount",
				"responses": {
					"200": {
						"description": "Verified",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidationErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"name": "Validation Error.",
											"message": "Some fields are not valid.",
											"status": false,
											"errors": [
												{
													"message": "token is not allowed to be empty",
													"field": {
														"label": "token",
														"value": "",
														"key": "token"
													}
												}
											]
										}
									}
								}
							}
						}
					}
				},
				"tags": [
					"Authorization"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ITokenInputDTO"
							}
						}
					}
				}
			}
		},
		"/auth/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "LoggedIn",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LoggedUserResponse"
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidationErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"name": "Validation Error.",
											"message": "Some fields are not valid.",
											"status": false,
											"errors": [
												{
													"message": "email is not allowed to be empty",
													"field": {
														"label": "email",
														"value": "",
														"key": "email"
													}
												}
											]
										}
									}
								}
							}
						}
					}
				},
				"tags": [
					"Authorization"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUserLoginDTO"
							}
						}
					}
				}
			}
		},
		"/auth/refresh-token": {
			"post": {
				"operationId": "RefreshToken",
				"responses": {
					"200": {
						"description": "Refreshed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LoggedUserResponse"
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidationErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"name": "Validation Error.",
											"message": "Some fields are not valid.",
											"status": false,
											"errors": [
												{
													"message": "token is not allowed to be empty",
													"field": {
														"label": "token",
														"value": "",
														"key": "token"
													}
												}
											]
										}
									}
								}
							}
						}
					}
				},
				"description": "Use a refresh token to generate a new JWT token and a new refresh token",
				"tags": [
					"Authorization"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ITokenInputDTO"
							}
						}
					}
				}
			}
		},
		"/auth/revoke-token": {
			"post": {
				"operationId": "RevokeToken",
				"responses": {
					"200": {
						"description": "Revoked",
						"content": {
							"application/json": {
								"schema": {
									"type": "boolean"
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidationErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"name": "Validation Error.",
											"message": "Some fields are not valid.",
											"status": false,
											"errors": [
												{
													"message": "token is not allowed to be empty",
													"field": {
														"label": "token",
														"value": "",
														"key": "token"
													}
												}
											]
										}
									}
								}
							}
						}
					}
				},
				"tags": [
					"Authorization"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ITokenInputDTO"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/api/v1"
		}
	]
}